% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_data.R
\name{latest_data}
\alias{latest_data}
\title{Get latest data from Washington Post Github Repo}
\usage{
latest_data()
}
\value{
tibble
}
\description{
Checks if new data is available from the Washington Post Github repo and returns the most current dataset.
}
\examples{
latest_data()
}
