% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waposchosh-package.R
\docType{package}
\name{waposchosh-package}
\alias{waposchosh}
\alias{waposchosh-package}
\title{waposchosh: R package containing Washington Post's school shooting dataset}
\description{
Provides Washington Post Police Shootings dataset. Data source: Washington Post Github \url{https://github.com/washingtonpost/data-school-shootings}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/parmsam/waposchosh}
  \item \url{https://parmsam.github.io/waposchosh/}
  \item Report bugs at \url{https://github.com/parmsam/waposchosh/issues}
}

}
\keyword{internal}
