% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waposhosh-package.R
\docType{package}
\name{waposhosh-package}
\alias{waposhosh}
\alias{waposhosh-package}
\title{waposhosh: R package containing Washington Post's school shooting dataset}
\description{
Provides Washington Post Police Shootings dataset. Data source: Washington Post Github \url{https://github.com/washingtonpost/data-school-shootings}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/parmsam/waposhosh}
  \item Report bugs at \url{https://github.com/parmsam/waposhosh/issues}
}

}
\keyword{internal}
